Homework8
A lottery ticket buyer purchases ten tickets a week, always playing the same ten 5-digit 
“lucky” combinations. Write a program that initializes an array or a vector with these 
numbers, then lets the player enter this week’s winning 5-digit number. The program hould perform a linear search through the list of the player’s numbers and report 
whether or not one of the tickets is a winner this week. Here are the numbers:
13579  26791  26792  33445  55555
62483  77777  79422  85647  93121