Add files via upload
Write a program that computes and displays the charges for a patient’s hospital 
stay. First, the program should ask if the patient was admitted as an inpatient or an 
outpatient. If the patient was an inpatient, the following data should be entered:
•	 The	number	of	days	spent	in	the	hospital
•	 The	daily	rate
•	 Hospital	medication	charges
•	 Charges	for	hospital	services	(lab	tests,	etc.)
The program should ask for the following data if the patient was an outpatient:
•	 Charges	for	hospital	services	(lab	tests,	etc.)
•	 Hospital	medication	charges
The program should use two overloaded functions to calculate the total charges. One 
of the functions should accept arguments for the inpatient data, while the other func-tion accepts arguments for outpatient information. Both functions should return the 
total charges.
Input Validation: Do not accept negative numbers for any data